+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod variable;
+pub mod variables;
 pub mod vector;
 
 pub mod prelude {}
