+++ src/variables/mod.rs
@@ -1,0 +1,3 @@
+pub mod scaling;
+pub mod traits;
+pub mod variableautoscale;
