+++ src/vector.rs
@@ -21,21 +21,6 @@ pub fn index_of_best_vector(m: ArrayView2<f64>) -> usize {
     l2_norm_vectors(m).argmin().unwrap()
 }
 
-pub fn rescale_vector(v: ArrayView1<f64>, shift: f64, scaling_factor: f64) -> Array1<f64> {
-    let shift_vector = Array1::from_elem(v.len(), shift);
-    (&v - shift_vector) * scaling_factor
-}
-
-pub fn rescale_and_invert_vector(
-    v: ArrayView1<f64>,
-    shift: f64,
-    scaling_factor: f64,
-) -> Array1<f64> {
-    // This is scaling and then a 180° rotation to ensure the min is now the max, and vice-versa.
-    let ones = Array1::ones(v.len());
-    (&rescale_vector(v, shift, scaling_factor) - ones) * -1.0
-}
-
 #[cfg(test)]
 mod tests {
     use ndarray::array;
@@ -94,51 +79,5 @@ mod tests {
         let m = array![[1., 1., 1.], [2., 0., 0.], [1., 0., 1.], [1., 0., 1.],];
         assert_eq!(l2_norm_vectors(m.view()), array![sqrt3, 2., SQRT2, SQRT2,]);
         assert_eq!(index_of_best_vector(m.view()), 2);
-    }
-
-    #[test]
-    fn rescale_vector_pos_already_scaled() {
-        let v = array![0., 0.5, 1.];
-        assert_eq!(rescale_vector(v.view(), 0., 1.), array![0., 0.5, 1.]);
-    }
-
-    #[test]
-    fn rescale_vector_pos_scaling() {
-        let v = array![1., 6., 11.];
-        let shift = v.min().unwrap();
-        let scaling_factor = 1.0 / (v.max().unwrap() - shift);
-        assert_eq!(
-            rescale_vector(v.view(), *shift, scaling_factor),
-            array![0., 0.5, 1.]
-        );
-    }
-
-    #[test]
-    fn rescale_vector_neg_scaling() {
-        let v = array![-3., -2., -1.];
-        let shift = v.min().unwrap();
-        let scaling_factor = 1.0 / (v.max().unwrap() - shift);
-        assert_eq!(
-            rescale_vector(v.view(), *shift, scaling_factor),
-            array![0., 0.5, 1.]
-        );
-    }
-
-    #[test]
-    fn rescale_vector_scaling() {
-        let v = array![0., 1., 6., 11., 12.];
-        assert_eq!(
-            rescale_vector(v.view(), 1., 0.1),
-            array![-0.1, 0., 0.5, 1., 1.1]
-        );
-    }
-
-    #[test]
-    fn rescale_and_invert_vector_full() {
-        let v = array![0., 1., 6., 11., 12.];
-        assert_eq!(
-            rescale_and_invert_vector(v.view(), 1., 0.1),
-            array![1.1, 1., 0.5, 0., -0.10000000000000009]
-        );
     }
 }
